Turn TypeOps into an object
Also, move `avoid` from TypeAssigner to TypeOps, since we
need to refer to it in Types.

Based on #8867
