+++ compiler/src/dotty/tools/dotc/core/TypeErasure.scala
@@ -5,6 +5,7 @@ package core
 import Symbols._, Types._, Contexts._, Flags._, Names._, StdNames._
 import Flags.JavaDefined
 import Uniques.unique
+import TypeOps.makePackageObjPrefixExplicit
 import transform.ExplicitOuter._
 import transform.ValueClasses._
 import transform.TypeUtils._
@@ -166,7 +167,7 @@ object TypeErasure {
   def erasedRef(tp: Type)(implicit ctx: Context): Type = tp match {
     case tp: TermRef =>
       assert(tp.symbol.exists, tp)
-      val tp1 = ctx.makePackageObjPrefixExplicit(tp)
+      val tp1 = makePackageObjPrefixExplicit(tp)
       if (tp1 ne tp) erasedRef(tp1)
       else TermRef(erasedRef(tp.prefix), tp.symbol.asTerm)
     case tp: ThisType =>
